+++ app/services/profile_matcher/lunarbaboon_profile_matcher.rb
@@ -1,0 +1,6 @@
+module ProfileMatcher
+  class LunarbaboonProfileMatcher < DomainMatcher
+    match_specificity 100
+    match_domains "lunarbaboon.com"
+  end
+end